v1.0 launched

+++ pages/api/products/[id].js
@@ -5,7 +5,9 @@ export default async function handler(req, res) {
     const {
         method,
         query: { id },
+        cookies,
     } = req;
+    const token = cookies.token;
 
     await dbConnect();
 
@@ -19,6 +21,9 @@ export default async function handler(req, res) {
     }
 
     if (method === 'PUT') {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json('You are not authenticated!');
+        }
         try {
             const product = await Product.find(req.body);
             res.status(201).json(product);
@@ -28,6 +33,9 @@ export default async function handler(req, res) {
     }
 
     if (method === 'DELETE') {
+        if (!token || token !== process.env.token) {
+            return res.status(401).json('You are not authenticated!');
+        }
         try {
             await Product.findByIdAndDelete(id);
             res.status(201).json(
